+++ src/packages/cli/src/__tests__/integration/mysql/runtime.test.ts
@@ -1,0 +1,5 @@
+import * as IntegrationTest from '../../__helpers__/integrationTest'
+import { database } from './__database'
+import { scenarios } from './__scenarios'
+
+IntegrationTest.runtimeIntegrationTest({ database, scenarios })
